research-paper-chatbot/requirements.txt
Flask==3.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0